Refactored the PIDController and control-related commands.
Refactored the PIDController to extract a reusable interface for any PID controller, and change the control-related command API and classes to distinguish between managed and unmanaged controllers.
